updated sidebar
