+++ app/javascript/react/components/App.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect } from 'react'
 import ComponentChoice from './ComponentCollection'
 import Sidebar from './Sidebar/Sidebar'
 
@@ -9,10 +9,30 @@ export const App = (props) => {
     id: null
   })
 
+  const [lists, setLists] = useState([])
+
+  const fetchLists = async() => {
+    try {
+      const response = await fetch("/api/v1/lists")
+      if (!response.ok) {
+        const errorMessage = `${response.status} (${response.statusText})`
+        throw new Error(errorMessage)
+      }
+      const listData = await response.json()
+      setLists(listData.lists)
+    } catch(error) {
+      console.error(`Error in fetch: ${error.message}`)
+    }
+  }
+
+  useEffect(() => {
+    fetchLists()
+  }, [])
+
   return (
     <>
-      <Sidebar />
-      <ComponentChoice page={page} setPage={setPage} />
+      <Sidebar lists={lists} setPage={setPage}/>
+      <ComponentChoice page={page} lists={lists} setPage={setPage} />
     </>
   )
 }
